third_party/WebKit/LayoutTests/virtual/gpu/fast/canvas/webgl/oes-vertex-array-object-expected.txt

CONSOLE MESSAGE: line 24: *** Error loading shader '[object WebGLShader]':INVALID_OPERATION
CONSOLE MESSAGE: line 24: *** Error loading shader '[object WebGLShader]':INVALID_VALUE
CONSOLE MESSAGE: line 24: Error in program linking:missing shaders
This test verifies the functionality of the OES_vertex_array_object extension, if it is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists
Testing binding enum with extension disabled
PASS getError was expected value: INVALID_ENUM : VERTEX_ARRAY_BINDING_OES should not be queryable if extension is disabled
PASS Successfully enabled OES_vertex_array_object extension
PASS OES_vertex_array_object listed as supported and getExtension succeeded
Testing binding enum with extension enabled
PASS ext.VERTEX_ARRAY_BINDING_OES is 0x85B5
PASS getError was expected value: NO_ERROR : VERTEX_ARRAY_BINDING_OES query should succeed if extension is enable
PASS Default value of VERTEX_ARRAY_BINDING_OES is null
Testing binding a VAO
PASS gl.getParameter(ext.VERTEX_ARRAY_BINDING_OES) is null
FAIL gl.getParameter(ext.VERTEX_ARRAY_BINDING_OES) is not expected VAO
FAIL gl.getParameter(ext.VERTEX_ARRAY_BINDING_OES) is not expected VAO
PASS gl.getParameter(ext.VERTEX_ARRAY_BINDING_OES) is null
PASS getError was expected value: INVALID_OPERATION : binding a deleted vertex array object
PASS gl.getParameter(ext.VERTEX_ARRAY_BINDING_OES) is null
Testing object creation
PASS getError was expected value: NO_ERROR : createVertexArrayOES should not set an error
PASS vao is non-null.
PASS ext.isVertexArrayOES(vao) is false
FAIL ext.isVertexArrayOES(vao) should be true. Was false.
FAIL ext.isVertexArrayOES(vao) should be true. Was false.
PASS ext.isVertexArrayOES() is false
PASS ext.isVertexArrayOES(null) is false
Testing attributes work across bindings
FAIL ELEMENT_ARRAY_BUFFER_BINDING not preserved
FAIL ELEMENT_ARRAY_BUFFER_BINDING not preserved
FAIL ELEMENT_ARRAY_BUFFER_BINDING not preserved
FAIL ELEMENT_ARRAY_BUFFER_BINDING not preserved
FAIL ELEMENT_ARRAY_BUFFER_BINDING not preserved
FAIL ELEMENT_ARRAY_BUFFER_BINDING not preserved
FAIL ELEMENT_ARRAY_BUFFER_BINDING not preserved
FAIL ELEMENT_ARRAY_BUFFER_BINDING not preserved
Testing that attribute values are not attached to bindings
PASS Vertex attribute values are not attached to bindings
Testing draws with various VAO bindings
FAIL Draw 0 failed pixel test
FAIL Draw 1 failed pixel test
PASS Draw 2 passed pixel test
Testing using deleted buffers referenced by VAOs
FAIL buffer removed too early
FAIL buffer removed too early
FAIL buffer removed too early
FAIL should be 255,0,0,255
at (0, 0) expected: 255,0,0,255 was 255,255,255,255
FAIL should be 0,255,0,255
at (0, 0) expected: 0,255,0,255 was 255,255,255,255
FAIL should be 0,0,255,255
at (0, 0) expected: 0,0,255,255 was 255,255,255,255
FAIL should be 0,255,255,255
at (0, 0) expected: 0,255,255,255 was 255,255,255,255
Testing that VAOs don't effect ARRAY_BUFFER binding.
FAIL should be red
at (0, 0) expected: 255,0,0,255 was 255,255,255,255
FAIL should be green
at (0, 0) expected: 0,255,0,255 was 255,255,255,255
FAIL getError expected: NO_ERROR. Was INVALID_VALUE : there should be no errors

PASS successfullyParsed is true

TEST COMPLETE
